\name{ErccDbLite.ERCC.97}
\docType{package}

\alias{ErccDbLite.ERCC.97-package}
\alias{ErccDbLite.ERCC.97}
\alias{ErccDbLite.ERCC.97}


\title{ERCC annotation package}

\description{
  This package loads an SQL connection to a database containing
  annotations from Ensembl. For examples and help on functions see the
  help pages from the \code{ensembldb} package!
}

\note{
  This data package was made from resources at ERCC/NIST on
  Mon Apr 25 17:51:38 2016 and based on the N/A
}

\author{.}



\examples{
## load the library
##library(ErccDbLite.ERCC.97)
## list the contents that are loaded into memory
ls('package:ErccDbLite.ERCC.97')
## show the db object that is loaded by calling it's name
ErccDbLite.ERCC.97

## for more examples see the ensembldb package.


}

\keyword{package}
\keyword{data}
